+++ example/lib/main.dart
@@ -4,8 +4,8 @@ import 'package:pytorch_lite/pytorch_lite.dart';
 import 'package:pytorch_lite_example/run_model_by_camera_demo.dart';
 import 'package:pytorch_lite_example/run_model_by_image_demo.dart';
 
-void main() {
-  PytorchFfi.init();
+Future<void> main() async {
+  await PytorchFfi.init();
   runApp(const ChooseDemo());
 }